@@ -20,59 +24,43 @@ keywords = [
     "ligand-receptor",
     "interaction",
 ]
-readme = "README.rst"
-include = ["LICENSE"]
-
-[tool.poetry.dependencies]
-python = ">=3.10"
-pandas = ">=1.5,<3.0"
-numpy = "<=2.2.1"
-plotnine = "<=0.12.4"
-seaborn = ">=0.12,<0.14"
-requests = ">=2.28.0"
-python-circos = ">=0.3.0"
-setuptools = "*"
-anndata = { optional = true, version = ">=0.7.6,<0.11.2" }
-black = { optional = true, version = "*" }
-pytest-cov = { optional = true, version = "*" }
-sphinx = { optional = true, version = "<9" }                        # see issue at https://github.com/readthedocs/readthedocs.org/issues/10279
-nbsphinx = { optional = true, version = "<=0.9.5" }
-sphinx-autodoc-typehints = { optional = true, version = "<=2.5.0" }
-sphinx_rtd_theme = { optional = true, version = "<=3.0.2" }
-readthedocs-sphinx-ext = { optional = true, version = "<=2.2.5" }
-recommonmark = { optional = true, version = "<=0.7.1" }
-
-[tool.poetry.dev-dependencies]
-
-
-[tool.poetry.extras]
-test = ["anndata", "black", "pytest-cov"]
-docs = [
-    "sphinx",
-    "nbsphinx",
-    "sphinx-autodoc-typehints",
-    "sphinx_rtd_theme",
-    "readthedocs-sphinx-ext",
-    "recommonmark",
+readme = { file = "README.rst", content-type = "text/x-rst" }
+requires-python = ">=3.10"
+dependencies = [
+    "pandas>=1.5,<3.0",
+    "numpy<=2.1.3",
+    "plotnine>=0.13.0",
+    "seaborn>=0.12,<0.14",
+    "requests>=2.28.0",
+    "python-circos",
+    # "biopython @ https://github.com/biopython/biopython.git",
 ]
 
-[build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
+[project.urls]
+homepage = "https://github.com/zktuong/ktplotspy"
+repository = "https://github.com/zktuong/ktplotspy"
 
-# Configuration for pytest
+[project.optional-dependencies]
+test = ["anndata>=0.7.6,<0.11.2", "black", "pytest-cov"]
+docs = [
+    "scanpy",
+    "ipykernel",
+    "sphinx<8",
+    "nbsphinx<=0.9.5",
+    "sphinx-autodoc-typehints<=2.3.0",
+    "sphinx_rtd_theme<=3.0.2",
+    "readthedocs-sphinx-ext<=2.2.5",
+    "recommonmark<=0.7.1",
+]
+
 [tool.pytest.ini_options]
 filterwarnings = [
     "ignore::DeprecationWarning",
     "ignore::plotnine.exceptions.PlotnineWarning",
 ]
 testpaths = ["tests"]
-addopts = """\
-    --cov ktplotspy \
-    --cov-report=xml
-"""
+addopts = "--cov src/ktplotspy --cov-report=xml"
 
-# Configuration for the interrogate tool which checks docstring coverage
 [tool.interrogate]
 ignore-init-method = true
 ignore-init-module = true
@@ -92,3 +80,42 @@ exclude = ["test", "docs"]
 [tool.black]
 line-length = 140
 include = '\.pyi?$'
+
+[tool.setuptools.packages.find]
+where = ["src"]
+
+[tool.setuptools]
+include-package-data = false
+
+[tool.setuptools_scm]
+version_scheme = "guess-next-dev"
+local_scheme = "dirty-tag"
+
+[tool.pixi.project]
+channels = ["conda-forge"]
+platforms = ["osx-arm64", "linux-64", "win-64"]
+
+[tool.pixi.pypi-dependencies]
+ktplotspy = { path = ".", editable = true }
+
+[tool.pixi.environments]
+default = ["py310"]
+docs = { features = ["py312", "docs"], solve-group = "py312" }
+test = { features = ["py312", "test"], solve-group = "py312" }
+py310 = ["py310", "docs", "test"]
+py311 = ["py311", "docs", "test"]
+py312 = ["py312", "docs", "test"]
+# py313 = ["py313", "docs", "test"] # until biopython is updated to >=1.85 on pypi
+
+[tool.pixi.tasks]
+test = "pytest"
+coverage = "coverage report"
+
+[tool.pixi.feature.py310.dependencies]
+python = "~=3.10.0"
+[tool.pixi.feature.py311.dependencies]
+python = "~=3.11.0"
+[tool.pixi.feature.py312.dependencies]
+python = "~=3.12.0"
+# [tool.pixi.feature.py313.dependencies]
+# python = "~=3.13.0"
